This is a failing test for the cause of the issue mentioned in https://github.com/aws/aws-sdk-go/issues/457 and https://github.com/jmespath/go-jmespath/issues/14.

I've added a potential implementation that fixes the tests, but I'm not super happy with having to type switch against all the basic scalars in Go.