@ejholmes 2016-01-05
Didn't run the benchmarks actually. I can do a comparison before and after, later.