@jasdel 2016-01-19
I ran a benchmark locally with this change and the tip of master. Overall it looks like a small improvement in performance. I also ran this with `-benchmem` and see zero change in memory usage.

With nested structs having the largest negative impact.

```
benchmark                                   old ns/op     new ns/op     delta
BenchmarkInterpretSingleFieldStruct-8       315           298           -5.40%
BenchmarkInterpretNestedStruct-8            1082          1034          -4.44%
BenchmarkInterpretNestedMaps-8              166           158           -4.82%
BenchmarkLexIdentifier-8                    635           589           -7.24%
BenchmarkLexSubexpression-8                 1179          1190          +0.93%
BenchmarkLexDeeplyNested50-8                10355         10091         -2.55%
BenchmarkLexDeepNested50Pipe-8              12017         11246         -6.42%
BenchmarkLexDeepNested50Index-8             16917         16173         -4.40%
BenchmarkLexQuotedIdentifier-8              2518          2392          -5.00%
BenchmarkLexQuotedIdentifierEscapes-8       2997          2896          -3.37%
BenchmarkLexRawStringLiteral-8              1601          1550          -3.19%
BenchmarkLexDeepProjection104-8             45501         44718         -1.72%
BenchmarkParseIdentifier-8                  800           794           -0.75%
BenchmarkParseSubexpression-8               1648          1642          -0.36%
BenchmarkParseDeeplyNested50-8              24430         23938         -2.01%
BenchmarkParseDeepNested50Pipe-8            23981         24644         +2.76%
BenchmarkParseDeepNested50Index-8           29413         29203         -0.71%
BenchmarkParseQuotedIdentifier-8            2761          2740          -0.76%
BenchmarkParseQuotedIdentifierEscapes-8     3181          3260          +2.48%
BenchmarkParseRawStringLiteral-8            1846          1867          +1.14%
BenchmarkParseDeepProjection104-8           99096         99324         +0.23%
```