@jamesls 2016-01-19
@jasdel Thanks.  Does this overall approach look good to you?