@jasdel 2016-01-29
The change looks good for this test case which is great and it does fix this bug with `jpgo`, but I don't think it addresses the issue found in aws/aws-sdk-go#457. We'll need to fix that separately.