@ejholmes 2016-01-29
@jasdel This specifically solves the issue mentioned in that PR.