@jamesls 2016-02-01
Alright, I'll go ahead and merge this in, and then pick up any additional work.