@jamesls 2016-02-02
Looks like there was an unintended side effect of this pull request mentioned in #20.  I'm going to have to revert this commit.  As an alternative, it looks like #18 fixes the original issue as well.